+++ samples/multiple-files-project/build.sh
@@ -1,0 +1,1 @@
+g++ -Wno-write-strings main.cpp vehicle.cpp -o bin/main.cpp.bin