+++ samples/qa_logs.cpp
@@ -4,6 +4,9 @@
  * @author mkhan3189
  */
 #include "../easylogging++.h"
+
+_INITIALIZE_EASYLOGGINGPP
+
 int main(int argc, char** argv) {
   INFO("This is demo for QA logs");
   QA("I will only be printed if _QUALITY_ASSURANCE is defined");